fix: 休暇申請・有給残高まわりの不具合修正と未実装機能の追加 (closes #1-#12)
- 有給残高チェック追加（不足時は申請拒否・全ロールバック）(#1)
- 期間指定休暇申請の締め月ロック検証を追加 (#2)
- 半休の0.5日減算・返還・集計対応（単日制限バリデーション付き）(#3)
- 休暇申請の期間重複チェック追加＋期間重複SQLの判定修正 (#4)
- 休暇申請フォーム画面（leave-create.html）を本実装 (#5)
- 月次勤怠CSVエクスポート GET /attendance/export を追加 (#6)
- 有給一括付与の手動バッチ（当年付与済みスキップ・件数表示）を追加 (#7)
- 有給残日数の算出を全画面で paid_leave_balance 集計に統一 (#8)
- 36協定閾値をシステム設定のアラート閾値から取得するよう変更 (#9)
- 一括保存・承認一覧のN+1クエリを解消 (#10)
- 勤怠管理画面のキーワード検索を社員番号(empNo)照合に修正 (#11)
- 休暇申請一覧画面（leave-list.html）を本実装（取消・フィルタ対応）(#12)

ユニットテスト353件成功。PC/モバイル両表示の実機確認済み。

+++ src/main/java/com/attendance/app/mapper/LeaveApplicationMapper.java
@@ -4,6 +4,7 @@ import com.attendance.app.entity.LeaveApplication;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -110,11 +111,12 @@ public interface LeaveApplicationMapper {
     int reject(@Param("applicationId") Long applicationId);
 
     /**
-     * 指定年の承認済み有給使用日数を集計（ユーザー別）
+     * 指定年の承認済み有給使用日数を集計（ユーザー別）。
+     * 半休（AM_HALF/PM_HALF）は0.5日として集計されます。
      *
      * @param userId ユーザーID
      * @param year 年
      * @return 使用日数
      */
-    long countApprovedPaidLeaveDays(@Param("userId") Long userId, @Param("year") int year);
+    BigDecimal countApprovedPaidLeaveDays(@Param("userId") Long userId, @Param("year") int year);
 }
